feat(kilo-chat-hooks): create shared package; extract useConversations
