+++ apps/web/package.json
@@ -45,6 +45,7 @@
     "@kilocode/encryption": "workspace:*",
     "@kilocode/event-service": "workspace:*",
     "@kilocode/kilo-chat": "workspace:*",
+    "@kilocode/kilo-chat-hooks": "workspace:*",
     "@kilocode/kiloclaw-secret-catalog": "workspace:*",
     "@kilocode/worker-utils": "workspace:*",
     "@lottiefiles/dotlottie-react": "^0.17.15",
